back/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
Deprecated==1.2.18
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.0
Flask-Limiter==3.12
fsspec==2025.5.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
limits==5.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
mysql-connector-python==9.3.0
mysqlclient==2.2.7
networkx==3.5
numpy==2.2.6
nvidia-cufile-cu12==1.11.1.6
nvidia-cusparselt-cu12==0.6.3
nvidia-nvjitlink-cu12==12.6.85
ordered-set==4.1.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.1
PyJWT==2.8.0
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.3
sympy==1.14.0
threadpoolctl==3.6.0
torchaudio==2.7.0
torchvision==0.22.0
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
